fix: resolve type mismatches and test failures
- Fix PostgreSQL type mismatch in airport fuzzy search (Float8 -> Float4)
  SIMILARITY() returns REAL (4 bytes) not DOUBLE PRECISION (8 bytes)
- Add fallback to empty array when loading flights to prevent undefined errors
- Ensure flights is always an array even when response is malformed
- Fix aprs_filters test: Area filter is at index 5, not 4
- Fix round_trip test: accept floating point format without .0 suffix

+++ src/airports_repo.rs
@@ -427,9 +427,9 @@ impl AirportsRepository {
                 wikipedia_link: Option<String>,
                 #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::Text>)]
                 keywords: Option<String>,
-                #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::Float8>)]
+                #[diesel(sql_type = diesel::sql_types::Nullable<diesel::sql_types::Float4>)]
                 #[allow(dead_code)]
-                similarity_score: Option<f64>,
+                similarity_score: Option<f32>,
             }
 
             let results: Vec<AirportWithSimilarity> = diesel::sql_query(sql)
